requirements.txt
python-telegram-bot==21.4
openai>=1.20.0
sqlalchemy>=2.0.0
pydantic>=2.7.0
asyncpg>=0.29.0
langchain>=0.2.0
chromadb>=0.4.24
python-dotenv>=1.0.1
certifi>=2024.7.4
redis>=5.0.0
aiohttp>=3.9.5
pgvector>=0.2.4
supabase>=2.6.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
firecrawl>=4.0.0
pypdf >= 6.0.0
python-docx >= 1.2.0
docx2txt >= 0.8
Pillow >= 11.0.0